requirements.txt
torch==1.13.0
torchvision==0.14.0
facenet-pytorch==2.5.2
opencv-python==4.6.0.66
numpy==1.23.1
torch-directml==0.6.0
Pillow==10.0.1